combined functionality of login and sign up page
