+++ controllers/api/index.js
@@ -2,7 +2,9 @@
 
 const router = require('express').Router();
 const userRoutes = require('./userRoutes');
+// const gossipRoutes = require('./gossipRoutes');
 
 router.use('/users', userRoutes);
+// router.use('/gossip', gossipRoutes);
 
 module.exports = router;
