+++ controllers/api/userRoutes.js
@@ -1,13 +1,17 @@
 const router = require('express').Router();
 const { User } = require('../../models');
 
+
 router.post('/', async (req, res) => {
   try {
-    const dbUserData = await User.create(req.body);
+    const dbUserData = await User.create({
+      name: req.body.name,
+      email: req.body.email,
+      password: req.body.password,
+    });
 
     // Set up sessions with a 'loggedIn' variable set to `true`
     req.session.save(() => {
-      req.session.user_id = dbUserData.id;
       req.session.loggedIn = true;
 
       res.status(200).json(dbUserData);
@@ -17,47 +21,6 @@ router.post('/', async (req, res) => {
     res.status(500).json(err);
   }
 });
-
-router.post('/login', async (req, res) => {
-  try {
-    const { email, password } = req.body;
-
-    // Check if the provided credentials match the admin's credentials
-    const isdbUser = email === process.env.USER_EMAIL && password === process.env.USER_PASSWORD;
-
-    let userData;
-
-    if (isdbUser) {
-      // If the user is an admin, set the user data accordingly
-      userData = {
-        id: 1, 
-        name: 'User',
-        email: process.env.USER_EMAIL,
-        isUser: true,
-      };
-    } else {
-      // If not admin, proceed with regular user authentication
-      userData = await User.findOne({ where: { email: email } });
-
-      if (!userData || !(await userData.checkPassword(password))) {
-        // Invalid credentials
-        res.status(400).json({ message: 'Incorrect email or password, please try again' });
-        return;
-      }
-    }
-
-    // Create session variables based on the logged-in user
-    req.session.userId = userData.id;
-    req.session.loggedIn = true;
-    
-    // Send a JSON response indicating successful login
-    res.json({ user: userData, message: 'You are now logged in!', redirectTo: '/' });
-
-  } catch (err) {
-    res.status(400).json(err);
-  }
-});
-
 
 router.post('/login', async (req, res) => {
   try {
@@ -100,7 +63,7 @@ router.post('/login', async (req, res) => {
 });
 
 router.post('/logout', (req, res) => {
-  if (req.session.logged_in) {
+  if (req.session.loggedIn) {
     // Remove the session variables
     req.session.destroy(() => {
       res.status(204).end();
@@ -109,6 +72,5 @@ router.post('/logout', (req, res) => {
     res.status(404).end();
   }
 });
-
 
 module.exports = router;
