+++ controllers/homeRoutes.js
@@ -1,107 +1,33 @@
 
 const router = require('express').Router();
-const { Gossip, User } = require('../models');
+const { User } = require('../models');
 const withAuth = require('../utils/auth');
 
 router.get('/', async (req, res) => {
-  try {
-    const userData = await User.findAll({
-        include: [
-          {
-            model: User,
-          attributes: ['name']
-          },
-        ],
-       });
-
-      const users = userData.map((user) => user.get({ plain: true }));
-
-      res.render('homepage', {
-        users,
-        // Pass the logged in flag to the template
-        logged_in: req.session.logged_in,
-      });
+    try {
+        const userData = await User.findAll({
+            attributes: { exclude: ['password'] },
+            order: [['name', 'ASC']],
+        });
+        
+        const users = userData.map((project) => project.get({ plain: true }));
+        
+        res.render('homepage', {
+            users,
+            // Pass the logged in flag to the template
+            loggedIn: req.session.loggedIn,
+        });
     } catch (err) {
       res.status(500).json(err);
     }
   });
-
-router.get('/user/:id', async (req, res) => {
-  try {
-    const userData = await User.findByPk(req.params.id, {
-      include: [
-        {
-          model: Gossip,
-          attributes: ['name']
-        },
-      ],
-    });
-
-    const user = userData.get({ plain: true });
-
-    res.render('user', {
-      ...user,
-      logged_in: req.session.logged_in
-    });
-  } catch (err) {
-    res.status(500).json(err);
-  }
-});
-
-router.get('/gossip', async (req, res) => {
-  try {
-    // Get all projects and JOIN with user data
-    const gossipData = await Gossip.findAll({
-      include: [
-        {
-          model: User,
-          attributes: ['name'],
-        },
-      ],
-    });
-
-    // Serialize data so the template can read it
-    const gossip = gossipData.map((gossip) => gossip.get({ plain: true }));
-
-    // Pass serialized data and session flag into template
-    res.render('homepage', {
-      projects: gossip,
-      logged_in: req.session.logged_in
-    });
-  } catch (err) {
-    res.status(500).json(err);
-  }
-});
-
-router.get('/gossip/:id', async (req, res) => {
-  try {
-    const gossipData = await Gossip.findByPk(req.params.id, {
-      include: [
-        {
-          model: User,
-          attributes: ['name'],
-        },
-      ],
-    });
-
-    const gossip = gossipData.get({ plain: true });
-
-    res.render('gossip', {
-      ...gossip,
-      logged_in: req.session.logged_in
-    });
-  } catch (err) {
-    res.status(500).json(err);
-  }
-});
-
 
 router.get('/main', withAuth, async (req, res) => {
   try {
     // Render the 'main' view with any necessary data
     res.render('main', {
       // Pass data to the view if needed
-      logged_in: req.session.logged_in,
+      loggedIn: req.session.loggedIn,
     });
   } catch (err) {
     res.status(500).json(err);
@@ -109,20 +35,18 @@ router.get('/main', withAuth, async (req, res) => {
 });
 
 // Prevent non logged in users from viewing the dashboard
-router.get('/homepage', withAuth, async (req, res) => {
-  try {
+router.get('/dashboard', withAuth, async (req, res) => {
+    try {
     const userData = await User.findByPk(req.session.userId, {
-      attributes: { exclude: ['password'] },
-      include: [{ model: User }],
+      attributes: { exclude: ['password'] }
     });
 
     const user = userData.get({ plain: true });
-    res.render('homepage', {
-      ...user,
+    res.render('dashboard', {
+      user,
       // Pass the logged in flag to the template
-      logged_in: true,
-      // logged_in: req.session.logged_in,
-      // isAdmin: user.isAdmin,
+      loggedIn: req.session.loggedIn,
+      isAdmin: user.isAdmin,
     });
   } catch (err) {
     res.status(500).json(err);
@@ -131,8 +55,8 @@ router.get('/homepage', withAuth, async (req, res) => {
 
 router.get('/login', (req, res) => {
   // If a session exists, redirect the request to the homepage
-  if (req.session.logged_in) {
-    res.redirect('/homepage');
+  if (req.session.loggedIn) {
+    res.redirect('/');
     return;
   }
