+++ views/homepage.handlebars
@@ -16,4 +16,4 @@
         <button id= "signUpForm" class="column column-60">Submit</button>
     <div class="column column-20"> </div>
 </nav>
-   <script src="/js/sighnup.js"></script>
+   <script src="/js/signup.js"></script>
